requirements.txt
beautifulsoup4==4.5.1
bs4==0.0.1
certifi==2016.9.26
click==6.6
Flask==0.11.1
freeze==1.0.10
future==0.16.0
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-telegram-bot==5.2.0
pytz==2016.7
requests==2.11.1
requests-toolbelt==0.7.0
simplejson==3.10.0
six==1.10.0
telepot==10.1
urllib3==1.19.1
virtualenv==15.0.3
Werkzeug==0.11.11
yahoo-finance==1.3.2
google-api-python-client==1.7.3
